Add new admin  guis, improve commands

+++ plugin/src/main/java/dev/cerus/mapads/command/AdvertCommand.java
@@ -5,17 +5,21 @@ import co.aikar.commands.annotation.CommandAlias;
 import co.aikar.commands.annotation.CommandCompletion;
 import co.aikar.commands.annotation.CommandPermission;
 import co.aikar.commands.annotation.Dependency;
+import co.aikar.commands.annotation.Optional;
 import co.aikar.commands.annotation.Subcommand;
 import dev.cerus.mapads.advert.Advertisement;
 import dev.cerus.mapads.advert.storage.AdvertStorage;
+import dev.cerus.mapads.gui.AdDetailsGui;
+import dev.cerus.mapads.gui.AdListGui;
+import dev.cerus.mapads.gui.AdScreenListGui;
+import dev.cerus.mapads.gui.ConfirmationGui;
 import dev.cerus.mapads.image.storage.ImageStorage;
 import dev.cerus.mapads.lang.L10n;
-import java.util.List;
+import dev.cerus.mapads.screen.AdScreen;
+import dev.cerus.mapads.screen.storage.AdScreenStorage;
+import dev.cerus.mapads.util.ItemBuilder;
 import java.util.UUID;
-import net.md_5.bungee.api.chat.ClickEvent;
-import net.md_5.bungee.api.chat.ComponentBuilder;
-import net.md_5.bungee.api.chat.HoverEvent;
-import net.md_5.bungee.api.chat.hover.content.Text;
+import org.bukkit.Material;
 import org.bukkit.entity.Player;
 
 @CommandAlias("mapads")
@@ -29,17 +33,25 @@ public class AdvertCommand extends BaseCommand {
     @Dependency
     private ImageStorage imageStorage;
 
+    @Dependency
+    private AdScreenStorage adScreenStorage;
+
     @Subcommand("list")
     @CommandCompletion("@mapads_names")
-    public void handleList(final Player player, final String screenName) {
-        final List<Advertisement> list = this.advertStorage.getRunningAdvertisements(screenName);
-        player.sendMessage(L10n.get("prefix") + "§7" + list.size() + " ads");
-        list.forEach(advertisement -> {
-            player.spigot().sendMessage(new ComponentBuilder("§8- §f" + advertisement.getAdvertId().toString())
-                    .event(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new Text("§7Click to copy")))
-                    .event(new ClickEvent(ClickEvent.Action.COPY_TO_CLIPBOARD, advertisement.getAdvertId().toString()))
-                    .create());
-        });
+    public void handleList(final Player player, @Optional final String screenName) {
+        if (screenName == null) {
+            new AdScreenListGui(player, this.adScreenStorage, adScreen ->
+                    player.performCommand("mapads advert list " + adScreen.getId())).open();
+        } else {
+            final AdScreen adScreen = this.adScreenStorage.getAdScreen(screenName);
+            if (adScreen == null) {
+                player.sendMessage(L10n.getPrefixed("error.screen_not_found"));
+                return;
+            }
+
+            new AdListGui(this.advertStorage, adScreen, player, advertisement ->
+                    new AdDetailsGui(advertisement, player).open()).open();
+        }
     }
 
     @Subcommand("remove")
@@ -47,13 +59,21 @@ public class AdvertCommand extends BaseCommand {
     public void handleRemove(final Player player, final UUID advertId) {
         final Advertisement advert = this.advertStorage.getAdvert(advertId);
         if (advert == null) {
-            player.sendMessage(L10n.get("prefix") + "§cAdvert not found");
+            player.sendMessage(L10n.getPrefixed("error.advert_not_found"));
             return;
         }
 
-        this.advertStorage.deleteAdverts(advert.getAdvertId());
-        this.imageStorage.deleteMapImages(advert.getImageId());
-        player.sendMessage(L10n.get("prefix") + "§aAdvert was deleted");
+        new ConfirmationGui("Delete ad?", () ->
+                new ItemBuilder(Material.PAPER)
+                        .setName(L10n.get("command.advert.remove.confirm.name"))
+                        .setLore(L10n.getList("command.advert.remove.confirm.lore"))
+                        .build())
+                .onYes(() -> {
+                    player.closeInventory();
+                    this.advertStorage.deleteAdverts(advert.getAdvertId());
+                    this.imageStorage.deleteMapImages(advert.getImageId());
+                    player.sendMessage(L10n.getPrefixed("success.ad_deleted"));
+                }).onNo(player::closeInventory).open(player);
     }
 
 }
